Tratamento de exceção-DELETE

+++ src/main/java/com/educandoweb/course/services/UserServices.java
@@ -2,8 +2,10 @@ package com.educandoweb.course.services;
 
 import com.educandoweb.course.Repositories.UserRepository;
 import com.educandoweb.course.entities.User;
+import com.educandoweb.course.services.exceptions.DataBaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,13 +30,20 @@ public class UserServices {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        } catch (org.springframework.dao.DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
         updateData(entity, obj);
-        return repository.save(obj);
+        return repository.save(entity);
     }
 
     private void updateData(User entity, User obj) {
